Update link to Project Atlas
Project Atlas on Nexus updated to 0.7, but I don't have generated atlas textures for it based on the texture replacers on this list. And I really don't feel like doing the necessary work now (requires figuring out which textures need to be resized in Gimp). So for now I'm just linking to the old version 0.6.5.