Update Bloodmoon Rebalance PFP Patch
The patch is no longer available on Alvazir's Nexus page, so I uploaded a minimal version of it (with only records that still needed to be fixed in Bloodmoon Rebalance 1.2). Updated link, and added an explanation that the patch is still necessary even though Bloodmoon Rebalance includes some PFP fixes, because a few fixes are missed.